+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased][unreleased]
 ####  Added
+- added rename command
 - Added srem, sismember commands
 
 ### [1.0.0] - 2023-09-06